Fixede dit problem. Husk at bruge "add new item" når du tilføjer nye scripts
